Moved IMain ops requiring stability into implicit class.
A long-standing annoyance of having IMain stored in a var
is that you can't call a method on it which returns a dependent
type and then pass that to any other method.  I realized I
could get around this by creating an implicit class around
the var; in the class, it is a val, so the method can be written
there, and we implicitly convert from the var on demand.